requirements.txt
torch
torchvision
swanlab
transformers
diffusers
datasets
accelerate
clean-fid
pytorch_fid
openai-clip